Support for RPC
A callback can be registered to be called _only_ on receiving a response to a specific outgoing message.

I'm not entirely sure if this is the right API, vs. just registering a handler for a specific `_ok` message type and providing a set that just keys whether we sent this message and were expecting a response.

Still needs a working test.
May be easier to write an exchange in test files instead of writing string literals.