various API cleanups
The big one is that node no longer holds the creation time stdin/stdout.
Instead `run` controls that, which seems more logical since that is when
the ports are operated upon. It also reduces the number of fields in node.

The module also no longer re-exports stuff from maelstrom/message. Callers
should include the submodule.