3d solution
This is both stable under partitions and hits the criteria of <30msgs per op, median <400ms and max <600ms (in the non partition case).

The basic idea is taken from the maelstrom python solution of batching messages to peers. It builds on this by using a periodic timer instead of waiting for a response to handle dropped messages.

Then the topology is ignored. With the 100ms delay and the <400ms bound, no value should do more than 4 hops. This solution actually only does ~2 hops by having each node pick a subset of n/2+1 peers to broadcast to.
The n/2+1 ensures that each node has at least one peer that will keep it in the loop. I think. I need to think if this can be reduced.

I am still not clear about 3e because it seems to increase the times.