Simplify responding to a message
Got rid of `make-response`, which was _always_ used with `respond`.

Made `respond` take a hash as the argument instead of a assoc list, as
the latter requires escaping operations, and is less intuitive than just
using the actual form that will be serialized.

Get rid of the `current-input-msg` parameter, as it is prone to subtle
bugs where one run of the handler can capture the message, and if a sub
thread is spawned from the handler, and then used by another handler,
the captured message is incorrect.